Feature material

+++ server/public/scripts/services/actor.service.js
@@ -1,8 +1,11 @@
-app.service('ActorService', ['$http', function ($http) {
+app.service('ActorService', ['$http', function ($http, $mdDialog, $mdToast) {
     var self = this;
 
     self.actors = { data: '' };
 
+    self.newActor = { name: '' };
+
+    //Requests all actor objects from the database via the server and sends them to display on the DOM.
     self.getActors = function () {
         $http({
             method: 'GET',
@@ -17,6 +20,7 @@ app.service('ActorService', ['$http', function ($http) {
             });
     }
 
+    //Sends a new actor object to the server to get added to the database.
     self.addActor = function (newActor) {
         console.log('Add this Bond actor:', newActor);
         $http({
@@ -33,18 +37,21 @@ app.service('ActorService', ['$http', function ($http) {
             });
     }
 
+    //Requests deletion of an actor object from the database through the server.
     self.retireActor = function (actorId) {
-        console.log('retire actor', actorId);
-        $http({
-            method: 'DELETE',
-            url: `/actors/${actorId}`
-        }).then((response) => {
-            self.getActors();
-        }).catch((error) => {
-            console.log('error on /actors DELETE', error);
-        })
+        if (confirm('Are you sure you want to retire this Bond forever (forever...forever...)?')) {
+            console.log('retire actor', actorId);
+            $http({
+                method: 'DELETE',
+                url: `/actors/${actorId}`
+            }).then((response) => {
+                self.getActors();
+            }).catch((error) => {
+                console.log('error on /actors DELETE', error);
+            });
+        } else {
+            console.log('do NOT delete');
+        }
     }
-
-    self.getActors();
 
 }]);
